Create a parallel unit test for PETSc Multigrid
I had to make some changes to get this to work.  Most notably, I introduced a
level_info_local variable.  I also made it request level_info from the user by
default.  I realized that it is too difficult to try to generate level_info in
Futility.  It depends too strongly on the problem.

Also, I decided to go with PCSOR instead of trying block Jacobi.  Block Jacobi
only worked well in my MSED method when it was in red-black format.  PETSc
does not have support for red-black coloring, so block Jacobi doesn't work well.
PCSOR works pretty well so I'm making that the default option, but it might
not be as efficient.

Eventually, I want to write my own multigrid solver, but this will be a good
first implementation to get things rolling.