Change indexing of meshes in MultigridMeshStructureType
Using a 1-index to reduce confusion.  Originally, it was 0-indexed to line up
with PETSc but this should be easier to understand, especially if we want to
use this outside of PETSc.  Moreover, the coarsest grid (level 1) is
initialized now by default, but many of its parameters may not be initialized.