Add workaround for bug in PETSc with MatPtAP
The MatPtAP function results in an invalid read memory error in serial.  (It
tries to look for a nonlocal portion of the matrix even though there is only
one processor.)  The fix is to ensure that MatAIJ is used as the matrix type
instead of MatMPIAIJ when the # of local unknowns = the # of global unknowns
(i.e., the matrix is entirely conatined in a single processor).